changed file extensions and names to be less confusing
